+++ nes-core/include/Sinks/Mediums/MaterializedViewSink.hpp
@@ -72,7 +72,7 @@ class MaterializedViewSink : public SinkMedium {
       *  @brief Provides the id of the used materialized view
       *  @return materialized view id
       */
-     size_t getViewId() const;
+    uint64_t getViewId() const;
 
   private:
     MaterializedViewPtr view;
